Fix buggy BVH::world_bounds() impl
That should fix the weird lighting in the living_room scene as this
caused the world radius to be wrong. I also simplified the BVH
construction logic to be more in line with PBRT by taking in a vector of
Arc<Primitive> instead of Box, so we can properly build ordered_prim as
we go instead of having to mess with indices.